refactor(cli): make cli.py only responsible for parsing options and calling build refactor(api): move user-facing API (build()) and all related tools to yassg_tools.py
